requirements.txt
# requirements.txt for FinRL Research Project
# Generated based on requirements for core FinRL, RL algorithms,
# data handling, extensive visualization, and statistical/financial analysis.
# Current Date: 2025-05-05 (Use 'pip freeze > requirements.txt' for exact versions)

# -- Core FinRL --
# The main library
finrl>=0.3.5

# -- Reinforcement Learning Algorithms --
# Stable Baselines3 is commonly used by FinRL for DRL algorithms
# Ensure PyTorch is installed first or simultaneously if needed.
stable_baselines3[extra]>=2.0.0
# Use [extra] to include dependencies like Tensorboard logging

# Deep Learning Framework (Stable Baselines3 uses PyTorch)
# Check compatibility with stable-baselines3 version if specifying exact versions
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
# If using algorithms requiring TensorFlow (less common with recent SB3):
# tensorflow>=2.10.0

# Optional: Ray RLlib for different/distributed algorithms (heavier dependency)
# ray[rllib]>=2.5.0

# -- Data Handling and Manipulation --
# Essential for data structures and manipulation
pandas>=1.5.0
# Fundamental numerical computing library
numpy>=1.23.0
# Scientific computing library
scipy>=1.9.0

# -- Visualization --
# Standard plotting library
matplotlib>=3.6.0
# Higher-level statistical plotting based on Matplotlib
seaborn>=0.12.0
# Interactive plotting library (often used by FinRL's plot utilities)
plotly>=5.10.0
# Optional: Connects Plotly with Pandas for easy DataFrame plotting
# cufflinks>=0.17.3

# -- Statistical and Financial Analysis --
# Statistical models, tests, and time series analysis
statsmodels>=0.13.0
# General machine learning, includes preprocessing, metrics
scikit-learn>=1.1.0
# Comprehensive portfolio performance and risk analysis (maintained fork)
pyfolio-reloaded>=0.9.6
# Alternative/complementary library for portfolio quantitative analysis and reports
quantstats>=0.0.60

# -- Data Fetching (Commonly needed) --
# Popular library for downloading Yahoo Finance data
#yfinance>=0.2.10
# Optional: For WRDS database access (requires subscription)
# wrds>=3.1.0
# Optional: For Alpaca API (data and trading)
# alpaca-trade-api>=2.3.0
# Optional: For cryptocurrency exchange data
# ccxt>=3.0.0

# -- Development Environment (Recommended) --
# For interactive notebooks (essential for research workflows)
jupyter>=1.0.0
notebook>=6.5.0
jupyterlab>=3.5.0
# Allows Jupyter to use the project's virtual environment kernel
ipykernel>=6.15.0
# For interactive widgets in Jupyter
ipywidgets>=8.0.0

# -- Other Potential Utilities --
# Progress bars for loops
tqdm>=4.64.0
# TA-Lib wrapper (may require separate TA-Lib C library installation)
# TA-Lib>=0.4.24
# Stockstats is often used internally by FinRL for indicators
stockstats>=0.4.0

gymnasium~=1.1.1
PyYAML~=6.0.1
wandb~=0.19.10
ta~=0.11.0